+++ bzt/modules/apiritif/executor.py
@@ -16,21 +16,17 @@ limitations under the License.
 
 import os
 import re
-import shlex
 import sys
-import copy
 
-import yaml
 
 from bzt import TaurusConfigError
-from bzt.engine import HavingInstallableTools, SETTINGS
+from bzt.engine import SETTINGS
 from bzt.modules import SubprocessedExecutor, ConsolidatingAggregator, FuncSamplesReader
 from bzt.modules.functional import FunctionalResultsReader
 from bzt.modules.jmeter import JTLReader
-from bzt.six import string_types, text_type
-from bzt.utils import FileReader, get_full_path, RESOURCES_DIR, BZT_DIR
-from .generators import ApiritifScriptGenerator
-from .tools import TaurusPytestRunner, TaurusRobotRunner, Robot
+from bzt.utils import FileReader, get_full_path, RESOURCES_DIR, BZT_DIR, get_assembled_value
+from .generator import ApiritifScriptGenerator
+
 
 IGNORED_LINE = re.compile(r"[^,]+,Total:\d+ Passed:\d+ Failed:\d+")
 
@@ -83,7 +79,7 @@ class ApiritifNoseExecutor(SubprocessedExecutor):
         scenario = self.get_scenario()
 
         if test_mode == "apiritif":
-            builder = ApiritifScriptGenerator(self.engine, scenario, self.label, self.log, test_mode=test_mode)
+            builder = ApiritifScriptGenerator(scenario, self.label, test_mode=test_mode)
             builder.verbose = self.__is_verbose()
         else:
             wdlog = self.engine.create_artifact('webdriver', '.log')
@@ -91,24 +87,20 @@ class ApiritifNoseExecutor(SubprocessedExecutor):
             generate_markers = self.settings.get('generate-flow-markers', None)
             generate_markers = scenario.get('generate-flow-markers', generate_markers)
 
-            capabilities = copy.deepcopy(self.settings.get("capabilities"))
-            capabilities.merge(copy.deepcopy(self.execution.get("capabilities")))
-
-            scenario_caps = copy.deepcopy(scenario.get("capabilities"))
+            scenario_caps = scenario.get("capabilities")
 
             # todo: just for legacy support, remove it later
             if isinstance(scenario_caps, list):
                 self.log.warning("Obsolete format of capabilities found (list), should be dict")
-                scenario_caps = {item.keys()[0]: item.values()[0] for item in scenario_caps}
+                scenario["capabilities"] = {item.keys()[0]: item.values()[0] for item in scenario_caps}
 
-            capabilities.merge(scenario_caps)
+            configs = (self.settings, scenario, self.execution)
 
-            remote = self.settings.get("remote", None)
-            remote = self.execution.get("remote", remote)
-            remote = scenario.get("remote", remote)
+            capabilities = get_assembled_value(configs, "capabilities")
+            remote = get_assembled_value(configs, "remote")
 
             builder = ApiritifScriptGenerator(
-                self.engine, scenario, self.label, self.log, wdlog,
+                scenario, self.label, wdlog,
                 utils_file=os.path.join(RESOURCES_DIR, "selenium_taurus_extras.py"),
                 ignore_unknown_actions=self.settings.get("ignore-unknown-actions", False),
                 generate_markers=generate_markers,
@@ -256,151 +248,3 @@ class ApiritifFuncReader(FunctionalResultsReader):
                 if not self.read_records:
                     self.read_records = True
                 yield sample
-
-
-class PyTestExecutor(SubprocessedExecutor, HavingInstallableTools):
-    def __init__(self):
-        super(PyTestExecutor, self).__init__()
-        self.runner_path = os.path.join(RESOURCES_DIR, "pytest_runner.py")
-        self._tailer = FileReader('', file_opener=lambda _: None, parent_logger=self.log)
-        self._additional_args = []
-
-    def prepare(self):
-        super(PyTestExecutor, self).prepare()
-        self.install_required_tools()
-        self.script = self.get_script_path()
-        if not self.script:
-            raise TaurusConfigError("'script' should be present for pytest executor")
-
-        scenario = self.get_scenario()
-        if "additional-args" in scenario:
-            argv = scenario.get("additional-args")
-            self._additional_args = shlex.split(argv)
-
-        self.reporting_setup(suffix=".ldjson")
-
-    def __is_verbose(self):
-        engine_verbose = self.engine.config.get(SETTINGS).get("verbose", False)
-        executor_verbose = self.settings.get("verbose", engine_verbose)
-        return executor_verbose
-
-    def install_required_tools(self):
-        """
-        we need installed nose plugin
-        """
-        if sys.version >= '3':
-            self.log.warning("You are using Python 3, make sure that your scripts are able to run in Python 3")
-
-        self._check_tools([self._get_tool(TaurusPytestRunner, tool_path=self.runner_path)])
-
-    def startup(self):
-        """
-        run python tests
-        """
-        executable = self.settings.get("interpreter", sys.executable)
-
-        cmdline = [executable, self.runner_path, '--report-file', self.report_file]
-
-        load = self.get_load()
-        if load.iterations:
-            cmdline += ['-i', str(load.iterations)]
-
-        if load.hold:
-            cmdline += ['-d', str(load.hold)]
-
-        cmdline += self._additional_args
-        cmdline += [self.script]
-
-        self.process = self._execute(cmdline)
-
-        if self.__is_verbose():
-            self._tailer = FileReader(filename=self.stdout.name, parent_logger=self.log)
-
-    def check(self):
-        self.__log_lines()
-        return super(PyTestExecutor, self).check()
-
-    def post_process(self):
-        super(PyTestExecutor, self).post_process()
-        self.__log_lines()
-
-    def __log_lines(self):
-        lines = []
-        for line in self._tailer.get_lines():
-            if not IGNORED_LINE.match(line):
-                lines.append(line)
-
-        if lines:
-            self.log.info("\n".join(lines))
-
-
-class RobotExecutor(SubprocessedExecutor, HavingInstallableTools):
-    def __init__(self):
-        super(RobotExecutor, self).__init__()
-        self.runner_path = os.path.join(RESOURCES_DIR, "robot_runner.py")
-        self.variables_file = None
-        self.tags = None
-
-    def resource_files(self):
-        files = super(RobotExecutor, self).resource_files()
-        scenario = self.get_scenario()
-        if "variables" in scenario and isinstance(scenario["variables"], (string_types, text_type)):
-            files.append(scenario["variables"])
-        return files
-
-    def prepare(self):
-        super(RobotExecutor, self).prepare()
-        self.install_required_tools()
-        self.script = self.get_script_path()
-        if not self.script:
-            raise TaurusConfigError("'script' should be present for robot executor")
-
-        self.reporting_setup(suffix=".ldjson")
-
-        scenario = self.get_scenario()
-        variables = scenario.get("variables")
-        if variables:
-            if isinstance(variables, (string_types, text_type)):
-                self.variables_file = get_full_path(variables)
-            elif isinstance(variables, dict):
-                self.variables_file = self.engine.create_artifact("robot-vars", ".yaml")
-                with open(self.variables_file, 'wb') as fds:
-                    yml = yaml.safe_dump(variables,
-                                         default_flow_style=False, explicit_start=True, canonical=False,
-                                         allow_unicode=True,
-                                         encoding='utf-8', width=float("inf"))
-                    fds.write(yml)
-            else:
-                raise TaurusConfigError("`variables` is neither file nor dict")
-        tags = scenario.get("tags", None)
-        if tags:
-            if isinstance(tags, (string_types, text_type)):
-                self.tags = tags
-            else:
-                raise TaurusConfigError("`tags` is not a string or text")
-
-    def install_required_tools(self):
-        tools = [self._get_tool(TaurusRobotRunner, tool_path=self.runner_path),
-                 self._get_tool(Robot, python=self.settings.get("interpreter", sys.executable))]
-        self._check_tools(tools)
-
-    def startup(self):
-        executable = self.settings.get("interpreter", sys.executable)
-
-        cmdline = [executable, self.runner_path, '--report-file', self.report_file]
-
-        load = self.get_load()
-        if load.iterations:
-            cmdline += ['--iterations', str(load.iterations)]
-
-        if load.hold:
-            cmdline += ['--duration', str(load.hold)]
-
-        if self.variables_file is not None:
-            cmdline += ['--variablefile', self.variables_file]
-
-        if self.tags is not None:
-            cmdline += ['--include', self.tags]
-
-        cmdline += [self.script]
-        self.process = self._execute(cmdline)