+++ tests/resources/gatling/gatling3.sh
@@ -1,6 +1,7 @@
 #!/bin/sh
 
 COMPILER_CLASSPATH=
+GATLING_CLASSPATH=
 
 if [ -n "$GATLING_HOME" ]; then
   GATLING_DIR=${GATLING_HOME}/gatling
@@ -9,6 +10,6 @@ else
 fi
 
 JAVA="echo me"
-"$JAVA"
+bash -c "$JAVA"
 
 GATLING_HOME="$GATLING_HOME" COMPILATION_CLASSPATH="$COMPILATION_CLASSPATH" python -u ${GATLING_DIR}/gatling-fake.py $@